added stripe cli for testing

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -85,11 +85,13 @@ export class CheckoutPaymentComponent implements OnInit {
   async submitOrder() {
     this.loading = true;
     const basket = this.basketService.getCurrentBasketValue(); 
+    if(!basket) throw new Error('Cannot retrieve basket');
 
     try {
       const createdOrder = await this.createOrder(basket)
       const paymentResult = await this.confirmPaymentWithStripe(basket);
        if(paymentResult.paymentIntent){
+          this.basketService.deleteBasket(basket)
             this.basketService.deleteLocalBasket();
             const navigationExtras: NavigationExtras = {state: createdOrder};
             this.router.navigate(['/checkout/success'], navigationExtras);
